pyproject.toml
[project]
name = "eternaltwin"
version = "0.1.0"
description = "Python client for the Eternal-Twin API"
authors = [
    {name = "C. Quentin",email = "coumes.quentin@gmail.com"}
]
license = {text = "GNU Affero General Public License v3.0 or later <AGPL-3.0-or-later>"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)"
]

[dependency-groups]
dev = [
    "bandit (>=1.9.2,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "coverage (>=7.13.1,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pycodestyle (>=2.14.0,<3.0.0)",
    "pydocstyle (>=6.3.0,<7.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "types-requests (>=2.32.4.20260107,<3.0.0.0)",
    "pyflakes (>=3.4.0,<4.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

